AND-2419 Move spacers components to the UI core module
