+++ app/src/main/java/com/tangem/tap/common/compose/Warning.kt
@@ -13,6 +13,7 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.tangem.common.module.ModuleMessage
+import com.tangem.core.ui.components.SpacerH8
 import com.tangem.tap.domain.moduleMessage.ModuleMessageConverter
 import com.tangem.wallet.R
