+++ app/src/main/java/com/tangem/tap/features/home/compose/content/CurrenciesWeb3Content.kt
@@ -16,7 +16,7 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.DpSize
 import androidx.compose.ui.unit.dp
-import com.tangem.tap.common.compose.SpacerH
+import com.tangem.core.ui.components.SpacerH12
 import com.tangem.tap.common.compose.extensions.dpSize
 import com.tangem.tap.common.compose.extensions.halfHeight
 import com.tangem.tap.common.compose.extensions.toPx
@@ -64,7 +64,7 @@ fun StoriesCurrenciesContent(
                     targetOffset = animateTo.toPx(),
                     contentDescription = "Currency row",
                 )
-                SpacerH(12.dp)
+                SpacerH12()
             }
         }
     }
@@ -118,7 +118,7 @@ fun StoriesWeb3Content(
 
 @Composable
 private fun LightenBox(content: @Composable () -> Unit) {
-    Box() {
+    Box {
         content()
         Box(
             modifier = Modifier
@@ -130,10 +130,10 @@ private fun LightenBox(content: @Composable () -> Unit) {
                             Color.White.copy(alpha = 0f),
                             Color.White.copy(alpha = 0.75f),
                             Color.White.copy(alpha = 0.95f),
-                            Color.White
-                        )
-                    )
-                )
+                            Color.White,
+                        ),
+                    ),
+                ),
         ) {}
     }
 }
